app/backend/requirements.txt
--extra-index-url https://pkgs.dev.azure.com/azure-sdk/public/_packaging/azure-sdk-for-python/pypi/simple/
azure-identity==1.13.0b3
Flask==2.2.2
langchain==0.0.187
openai==0.26.4
azure-search-documents==11.4.0a20230509004
azure-storage-blob==12.14.1
tiktoken==0.4.0

